Split index into focused files
