+++ jest.config.js
@@ -5,7 +5,7 @@ module.exports = merge(config, {
   testEnvironment: 'jsdom',
   moduleFileExtensions: ['js', 'vue'],
   moduleNameMapper: {
-    '@testing-library/vue': '<rootDir>/src/vue-testing-library.js',
+    '@testing-library/vue': '<rootDir>/src/index.js',
   },
   coverageDirectory: './coverage',
   collectCoverageFrom: ['**/src/**/*.js', '!**/src/__tests__/**'],
